Change method definition of payment processors

+++ src/PaymentProcessor/StripePaymentProcessor.php
@@ -5,9 +5,10 @@ namespace Systemeio\TestForCandidates\PaymentProcessor;
 class StripePaymentProcessor
 {
     /**
+     * @param float $price payment amount in currency
      * @return bool true if payment was succeeded, false otherwise
      */
-    public function processPayment(int $price): bool
+    public function processPayment(float $price): bool
     {
         if ($price < 100) {
             return false;
